Move update_uimetadata function to kfp_utils
The function update_uimetadata is moved from jupyter_utils to kfp_utils,
it make much more sense to have a kfp specific function in kfp_utils.

All the referenced to the function in the tests are updated as well and
the unit_tests are moved to unit_tests/test_kfp_utils.py
