+++ kale/templates/function_template.jinja2
@@ -53,7 +53,8 @@ def {{ step_name }}({%- for arg in parameters_names -%}
 {% if in_variables|length > 0 or out_variables|length > 0 or function_body|length > 0 %}
     # run the code blocks inside a jupyter kernel
     from kale.utils.jupyter_utils import run_code as _kale_run_code
-    from kale.utils.jupyter_utils import update_uimetadata as _kale_update_uimetadata
+    from kale.utils.kfp_utils import \
+        update_uimetadata as _kale_update_uimetadata
     blocks = ({% if parameters_names|length > 0 %}pipeline_parameters_block,{% endif -%}
               {% if in_variables|length > 0 %}data_loading_block,{% endif -%}
 {%- for block in function_body %}
