+++ kale/tests/assets/functions/func04.out.py
@@ -10,7 +10,8 @@ def test():
 
     # run the code blocks inside a jupyter kernel
     from kale.utils.jupyter_utils import run_code as _kale_run_code
-    from kale.utils.jupyter_utils import update_uimetadata as _kale_update_uimetadata
+    from kale.utils.kfp_utils import \
+        update_uimetadata as _kale_update_uimetadata
     blocks = (data_loading_block,
               )
     html_artifact = _kale_run_code(blocks)
