+++ kale/tests/assets/kfp_dsl/pipeline_parameters_and_metrics.py
@@ -56,7 +56,8 @@ def create_matrix(d1: int, d2: int):
 
     # run the code blocks inside a jupyter kernel
     from kale.utils.jupyter_utils import run_code as _kale_run_code
-    from kale.utils.jupyter_utils import update_uimetadata as _kale_update_uimetadata
+    from kale.utils.kfp_utils import \
+        update_uimetadata as _kale_update_uimetadata
     blocks = (pipeline_parameters_block,
               block1,
               block2,
@@ -115,7 +116,8 @@ def sum_matrix():
 
     # run the code blocks inside a jupyter kernel
     from kale.utils.jupyter_utils import run_code as _kale_run_code
-    from kale.utils.jupyter_utils import update_uimetadata as _kale_update_uimetadata
+    from kale.utils.kfp_utils import \
+        update_uimetadata as _kale_update_uimetadata
     blocks = (data_loading_block,
               block1,
               block2,
