+++ kale/tests/unit_tests/test_jupyter_utils.py
@@ -12,11 +12,10 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import os
-import json
 import pytest
 
 from testfixtures import mock
+
 from kale.utils import jupyter_utils as ju
 
 
@@ -42,83 +41,6 @@ def _output_display(data):
 def test_generate_html_output(outputs, target):
     """Tests html artifact generation from cell outputs."""
     assert target == ju.generate_html_output(outputs)
-
-
-@mock.patch('kale.utils.jupyter_utils.pod_utils')
-def test_update_uimetadata_not_exists(pod_utils, tmpdir):
-    """Test the uimetadata file is created when it does not exists."""
-    pod_utils.get_pod_name.return_value = 'test_pod'
-    pod_utils.get_namespace.return_value = 'test_ns'
-    pod_utils.get_workflow_name.return_value = 'test_wk'
-
-    filepath = os.path.join(tmpdir, 'tmp_uimetadata.json')
-
-    # update tmp file
-    ju.update_uimetadata('test', uimetadata_path=filepath)
-
-    # check file has been updated correctly
-    updated = json.loads(open(filepath).read())
-    target = {"outputs": [{
-        'type': 'web-app',
-        'storage': 'minio',
-        'source': 'minio://mlpipeline/artifacts/test_wk/test_pod/test.tgz'
-    }]}
-    assert updated == target
-
-
-@mock.patch('kale.utils.jupyter_utils.pod_utils')
-def test_update_uimetadata_from_empty(pod_utils, tmpdir):
-    """Test that the uimetadata file is updated inplace correctly."""
-    pod_utils.get_pod_name.return_value = 'test_pod'
-    pod_utils.get_namespace.return_value = 'test_ns'
-    pod_utils.get_workflow_name.return_value = 'test_wk'
-
-    # create base tmp file
-    base = {"outputs": []}
-    filepath = os.path.join(tmpdir, 'tmp_uimetadata.json')
-    json.dump(base, open(filepath, 'w'))
-
-    # update tmp file
-    ju.update_uimetadata('test', uimetadata_path=filepath)
-
-    # check file has been updated correctly
-    updated = json.loads(open(filepath).read())
-    target = {"outputs": [{
-        'type': 'web-app',
-        'storage': 'minio',
-        'source': 'minio://mlpipeline/artifacts/test_wk/test_pod/test.tgz'
-    }]}
-    assert updated == target
-
-
-@mock.patch('kale.utils.jupyter_utils.pod_utils')
-def test_update_uimetadata_from_not_empty(pod_utils, tmpdir):
-    """Test that the uimetadata file is updated inplace correctly."""
-    pod_utils.get_pod_name.return_value = 'test_pod'
-    pod_utils.get_namespace.return_value = 'test_ns'
-    pod_utils.get_workflow_name.return_value = 'test_wk'
-
-    # create base tmp file
-    markdown = {
-        'type': 'markdown',
-        'storage': 'inline',
-        'source': '#Some markdown'
-    }
-    base = {"outputs": [markdown]}
-    filepath = os.path.join(tmpdir, 'tmp_uimetadata.json')
-    json.dump(base, open(filepath, 'w'))
-
-    # update tmp file
-    ju.update_uimetadata('test', uimetadata_path=filepath)
-
-    # check file has been updated correctly
-    updated = json.loads(open(filepath).read())
-    target = {"outputs": [markdown, {
-        'type': 'web-app',
-        'storage': 'minio',
-        'source': 'minio://mlpipeline/artifacts/test_wk/test_pod/test.tgz'
-    }]}
-    assert updated == target
 
 
 @mock.patch('kale.utils.jupyter_utils.process_outputs', new=lambda x: x)
